bot_v0.0.2/requirements.txt
aiogram==3.13.1
supabase==2.3.0
python-dotenv==1.0.0
aiohttp>=3.8.0
